feat: added tables to support incredible timetables in schema.prisma (#10)
* feat: added tables to support incredible timetables in schema.prisma

* feat: added a timetable service

* added prettier in .eslintrc
* updated .prettierrc
* updated dependencies
* removed startHour and endHour, replacing them by occurenceDuration
* updated typing for database models

* fix: timetable fetching should now work (not tested)

* refactor: refactored TimetableEntry structure

* overrides are now in a separate table
* we now use less times, and more metrics based on counting occurrences, which makes for easier maths and database consistency
* starting adding support for unit testing
* started testing the TimetableService class

* fix: somes bugs + more tests

* fix: lint

* fix: data cleaning was shlag

* fix: now pnpm test does unit and e2e tests + lint + unitSuite util function

* docs: documented TimetableService

* and fixed some bugs, oupsy
* updated scripts in package.json and added scripts test:unit and test:e2e
* lint

* feat: route GET /timetable/current/daily/:day/:month/:year

* removed dependency supertest
* timetable occurrences are now sorted

* feat: added utility function sortArray

* fix: changed parameter name day to date in route GET timetable/current/daily/:day/:month/:year

* new features

* fix: revert absolute paths

* fix previous tests

* fix tests for the first route (1/12)

* add more tests

* tests: add tests for criteria

* add interfaces, typings and selectors for ue types

* test: one more test

* test: more test

* test: forgot one test

* tests: add test for comment posting

* test: add more tests

* fix: add last tests

* removed all before alls

* tests: add tests to check for uuid format

* fix merge

* feat: added route GET /timetable/current/groups

* added field createdAt in schema.prisma
* added utility test function createTimetableGroup
* utility test function createUser now automatically fakes all data

* feat: added route GET /timetable/:entryId

* feat: created function create

* that's a draft atm

* test: fix test coverage

* add typing for @GetUser decorator
* add missing test
* add common typed create ue function for tests

* test: add shortcuts for Spec#expectJsonLike

* refactor: each entry (or override) can now be associated with multiple groups

* feat: added route POST /timetable/current

* feat: added route PATCH /timetable/current/:entryId

* fix: can no longer override an entry for groups that were not in the original entry

* experiment(test+errors): add custom errors on http 400 and 401

* add custom assertions to pactum. At this point, custom assertions only concern AppExceptions and UE-relative objects

* fix: add more details in custom http 400 errors

* test: end tests

* docs is still missing

* fix: tests

* feat: added route DELETE /timetable/current/:entryId

* refactor: some old tests that did not use current tools

* dependencies: upgrade + very small updates on old tests

* docs: added documentation

* fix: lint

* fix: lot of small fixes

* removed environment variables POSTGRES_USER, POSTGRES_PASSWORD, POSTGRES_DB, and added FAKER_SEED for test environment
* added file .env.test.dist for test environment
* updated commands in package.json
* changed User.RGPDId and User.RGPD to User.rgpdId and User.rgpd in schema.prisma
* now throwing AppException s instead of ForbiddenException s, NotFoundException s, ...
* removed useless @UseGuards(JwtGuard)
* removed default collect coverage when testing
* changed ES version

* feat: added a way to generate functions to create fake entities way more easily

* fix: review

* fix: added flags --runInBand in testing commands

---------

+++ src/ue/ue.controller.ts
@@ -15,7 +15,7 @@ import {
 import { UESearchDto } from './dto/ue-search.dto';
 import { UEService } from './ue.service';
 import { GetUser, IsPublic } from '../auth/decorator';
-import { User } from '../prisma/types';
+import { User } from '../users/interfaces/user.interface';
 import { UeCommentPostDto } from './dto/ue-comment-post.dto';
 import { AppException, ERROR_CODE } from '../exceptions';
 import { UERateDto } from './dto/ue-rate.dto';
@@ -36,40 +36,24 @@ export class UEController {
   @Get('/:ueCode')
   @IsPublic()
   async getUE(@Param('ueCode') ueCode: string) {
-    if (!(await this.ueService.doesUEExist(ueCode)))
-      throw new AppException(ERROR_CODE.NO_SUCH_UE, ueCode);
+    if (!(await this.ueService.doesUEExist(ueCode))) throw new AppException(ERROR_CODE.NO_SUCH_UE, ueCode);
     return this.ueService.getUE(ueCode.toUpperCase());
   }
 
   @Get('/:ueCode/comments')
-  async getUEComments(
-    @Param('ueCode') ueCode: string,
-    @GetUser() user: User,
-    @Query() dto: GetUECommentsDto,
-  ) {
-    if (!(await this.ueService.doesUEExist(ueCode)))
-      throw new AppException(ERROR_CODE.NO_SUCH_UE, ueCode);
-    return this.ueService.getComments(
-      ueCode,
-      user,
-      dto,
-      user.permissions.includes('commentModerator'),
-    );
+  async getUEComments(@Param('ueCode') ueCode: string, @GetUser() user: User, @Query() dto: GetUECommentsDto) {
+    if (!(await this.ueService.doesUEExist(ueCode))) throw new AppException(ERROR_CODE.NO_SUCH_UE, ueCode);
+    return this.ueService.getComments(ueCode, user.id, dto, user.permissions.includes('commentModerator'));
   }
 
   @Post('/:ueCode/comments')
-  async PostUEComment(
-    @Param('ueCode') ueCode: string,
-    @GetUser() user: User,
-    @Body() body: UeCommentPostDto,
-  ) {
-    if (!(await this.ueService.doesUEExist(ueCode)))
-      throw new AppException(ERROR_CODE.NO_SUCH_UE, ueCode);
-    if (!(await this.ueService.hasAlreadyDoneThisUE(user, ueCode)))
+  async PostUEComment(@Param('ueCode') ueCode: string, @GetUser() user: User, @Body() body: UeCommentPostDto) {
+    if (!(await this.ueService.doesUEExist(ueCode))) throw new AppException(ERROR_CODE.NO_SUCH_UE, ueCode);
+    if (!(await this.ueService.hasAlreadyDoneThisUE(user.id, ueCode)))
       throw new AppException(ERROR_CODE.NOT_ALREADY_DONE_UE);
-    if (await this.ueService.hasAlreadyPostedAComment(user, ueCode))
+    if (await this.ueService.hasAlreadyPostedAComment(user.id, ueCode))
       throw new AppException(ERROR_CODE.FORBIDDEN_ALREADY_COMMENTED);
-    return this.ueService.createComment(body, user, ueCode);
+    return this.ueService.createComment(body, user.id, ueCode);
   }
 
   @Patch('/comments/:commentId')
@@ -77,17 +61,16 @@ export class UEController {
     @Param(
       'commentId',
       new ParseUUIDPipe({
-        exceptionFactory: () => new AppException(ERROR_CODE.NOT_AN_UUID),
+        exceptionFactory: () => new AppException(ERROR_CODE.PARAM_NOT_UUID, 'commentId'),
       }),
     )
     commentId: string,
     @GetUser() user: User,
     @Body() body: UeCommentUpdateDto,
   ) {
-    if (!(await this.ueService.doesCommentExist(commentId)))
-      throw new AppException(ERROR_CODE.NO_SUCH_COMMENT);
-    if (await this.ueService.isUserCommentAuthor(user, commentId))
-      return this.ueService.updateComment(body, commentId, user);
+    if (!(await this.ueService.doesCommentExist(commentId))) throw new AppException(ERROR_CODE.NO_SUCH_COMMENT);
+    if (await this.ueService.isUserCommentAuthor(user.id, commentId))
+      return this.ueService.updateComment(body, commentId, user.id);
     throw new AppException(ERROR_CODE.NOT_COMMENT_AUTHOR);
   }
 
@@ -96,16 +79,15 @@ export class UEController {
     @Param(
       'commentId',
       new ParseUUIDPipe({
-        exceptionFactory: () => new AppException(ERROR_CODE.NOT_AN_UUID),
+        exceptionFactory: () => new AppException(ERROR_CODE.PARAM_NOT_UUID, 'commentId'),
       }),
     )
     commentId: string,
     @GetUser() user: User,
   ) {
-    if (!(await this.ueService.doesCommentExist(commentId)))
-      throw new AppException(ERROR_CODE.NO_SUCH_COMMENT);
-    if (await this.ueService.isUserCommentAuthor(user, commentId))
-      return this.ueService.deleteComment(commentId, user);
+    if (!(await this.ueService.doesCommentExist(commentId))) throw new AppException(ERROR_CODE.NO_SUCH_COMMENT);
+    if (await this.ueService.isUserCommentAuthor(user.id, commentId))
+      return this.ueService.deleteComment(commentId, user.id);
     throw new AppException(ERROR_CODE.NOT_COMMENT_AUTHOR);
   }
 
@@ -115,19 +97,18 @@ export class UEController {
     @Param(
       'commentId',
       new ParseUUIDPipe({
-        exceptionFactory: () => new AppException(ERROR_CODE.NOT_AN_UUID),
+        exceptionFactory: () => new AppException(ERROR_CODE.PARAM_NOT_UUID, 'commentId'),
       }),
     )
     commentId: string,
     @GetUser() user: User,
   ) {
-    if (!(await this.ueService.doesCommentExist(commentId)))
-      throw new AppException(ERROR_CODE.NO_SUCH_COMMENT);
-    if (await this.ueService.isUserCommentAuthor(user, commentId))
+    if (!(await this.ueService.doesCommentExist(commentId))) throw new AppException(ERROR_CODE.NO_SUCH_COMMENT);
+    if (await this.ueService.isUserCommentAuthor(user.id, commentId))
       throw new AppException(ERROR_CODE.IS_COMMENT_AUTHOR);
-    if (await this.ueService.hasAlreadyUpvoted(user, commentId))
+    if (await this.ueService.hasAlreadyUpvoted(user.id, commentId))
       throw new AppException(ERROR_CODE.FORBIDDEN_ALREADY_UPVOTED);
-    await this.ueService.upvoteComment(user, commentId);
+    await this.ueService.upvoteComment(user.id, commentId);
   }
 
   @Delete('/comments/:commentId/upvote')
@@ -136,19 +117,18 @@ export class UEController {
     @Param(
       'commentId',
       new ParseUUIDPipe({
-        exceptionFactory: () => new AppException(ERROR_CODE.NOT_AN_UUID),
+        exceptionFactory: () => new AppException(ERROR_CODE.PARAM_NOT_UUID, 'commentId'),
       }),
     )
     commentId: string,
     @GetUser() user: User,
   ) {
-    if (!(await this.ueService.doesCommentExist(commentId)))
-      throw new AppException(ERROR_CODE.NO_SUCH_COMMENT);
-    if (await this.ueService.isUserCommentAuthor(user, commentId))
+    if (!(await this.ueService.doesCommentExist(commentId))) throw new AppException(ERROR_CODE.NO_SUCH_COMMENT);
+    if (await this.ueService.isUserCommentAuthor(user.id, commentId))
       throw new AppException(ERROR_CODE.IS_COMMENT_AUTHOR);
-    if (!(await this.ueService.hasAlreadyUpvoted(user, commentId)))
+    if (!(await this.ueService.hasAlreadyUpvoted(user.id, commentId)))
       throw new AppException(ERROR_CODE.FORBIDDEN_ALREADY_UNUPVOTED);
-    await this.ueService.deUpvoteComment(user, commentId);
+    await this.ueService.deUpvoteComment(user.id, commentId);
   }
 
   @Post('/comments/:commentId/reply')
@@ -157,15 +137,14 @@ export class UEController {
     @Param(
       'commentId',
       new ParseUUIDPipe({
-        exceptionFactory: () => new AppException(ERROR_CODE.NOT_AN_UUID),
+        exceptionFactory: () => new AppException(ERROR_CODE.PARAM_NOT_UUID, 'commentId'),
       }),
     )
     commentId: string,
     @Body() body: CommentReplyDto,
   ) {
-    if (!(await this.ueService.doesCommentExist(commentId)))
-      throw new AppException(ERROR_CODE.NO_SUCH_COMMENT);
-    return this.ueService.replyComment(user, commentId, body);
+    if (!(await this.ueService.doesCommentExist(commentId))) throw new AppException(ERROR_CODE.NO_SUCH_COMMENT);
+    return this.ueService.replyComment(user.id, commentId, body);
   }
 
   @Patch('/comments/reply/:replyId')
@@ -174,16 +153,14 @@ export class UEController {
     @Param(
       'replyId',
       new ParseUUIDPipe({
-        exceptionFactory: () => new AppException(ERROR_CODE.NOT_AN_UUID),
+        exceptionFactory: () => new AppException(ERROR_CODE.PARAM_NOT_UUID, 'replyId'),
       }),
     )
     replyId: string,
     @Body() body: CommentReplyDto,
   ) {
-    if (!(await this.ueService.doesReplyExist(replyId)))
-      throw new AppException(ERROR_CODE.NO_SUCH_REPLY);
-    if (await this.ueService.isUserCommentReplyAuthor(user, replyId))
-      return this.ueService.editReply(replyId, body);
+    if (!(await this.ueService.doesReplyExist(replyId))) throw new AppException(ERROR_CODE.NO_SUCH_REPLY);
+    if (await this.ueService.isUserCommentReplyAuthor(user.id, replyId)) return this.ueService.editReply(replyId, body);
     throw new AppException(ERROR_CODE.NOT_REPLY_AUTHOR);
   }
 
@@ -193,15 +170,13 @@ export class UEController {
     @Param(
       'replyId',
       new ParseUUIDPipe({
-        exceptionFactory: () => new AppException(ERROR_CODE.NOT_AN_UUID),
+        exceptionFactory: () => new AppException(ERROR_CODE.PARAM_NOT_UUID, 'replyId'),
       }),
     )
     replyId: string,
   ) {
-    if (!(await this.ueService.doesReplyExist(replyId)))
-      throw new AppException(ERROR_CODE.NO_SUCH_REPLY);
-    if (await this.ueService.isUserCommentReplyAuthor(user, replyId))
-      return this.ueService.deleteReply(replyId);
+    if (!(await this.ueService.doesReplyExist(replyId))) throw new AppException(ERROR_CODE.NO_SUCH_REPLY);
+    if (await this.ueService.isUserCommentReplyAuthor(user.id, replyId)) return this.ueService.deleteReply(replyId);
     throw new AppException(ERROR_CODE.NOT_REPLY_AUTHOR);
   }
 
@@ -213,24 +188,17 @@ export class UEController {
 
   @Get('/:ueCode/rate')
   async GetRateUE(@Param('ueCode') ueCode: string, @GetUser() user: User) {
-    if (!(await this.ueService.doesUEExist(ueCode)))
-      throw new AppException(ERROR_CODE.NO_SUCH_UE, ueCode);
-    return this.ueService.getRateUE(user, ueCode);
+    if (!(await this.ueService.doesUEExist(ueCode))) throw new AppException(ERROR_CODE.NO_SUCH_UE, ueCode);
+    return this.ueService.getRateUE(user.id, ueCode);
   }
 
   @Put('/:ueCode/rate')
-  async RateUE(
-    @Param('ueCode') ueCode: string,
-    @GetUser() user: User,
-    @Body() dto: UERateDto,
-  ) {
-    if (!(await this.ueService.doesUEExist(ueCode)))
-      throw new AppException(ERROR_CODE.NO_SUCH_UE, ueCode);
-    if (!(await this.ueService.doesCriterionExist(dto.criterion)))
-      throw new AppException(ERROR_CODE.NO_SUCH_CRITERION);
-    if (!(await this.ueService.hasAlreadyDoneThisUE(user, ueCode)))
+  async RateUE(@Param('ueCode') ueCode: string, @GetUser() user: User, @Body() dto: UERateDto) {
+    if (!(await this.ueService.doesUEExist(ueCode))) throw new AppException(ERROR_CODE.NO_SUCH_UE, ueCode);
+    if (!(await this.ueService.doesCriterionExist(dto.criterion))) throw new AppException(ERROR_CODE.NO_SUCH_CRITERION);
+    if (!(await this.ueService.hasAlreadyDoneThisUE(user.id, ueCode)))
       throw new AppException(ERROR_CODE.NOT_ALREADY_DONE_UE);
-    return this.ueService.doRateUE(user, ueCode, dto);
+    return this.ueService.doRateUE(user.id, ueCode, dto);
   }
 
   /*
